+++ tasks/batch.sh
@@ -114,7 +114,7 @@ function batch() {
   date_header='Päivien '$start_day' - '$end_day' tehtävien priorisointi'
 
   # Save output to Obsidian vault with the start and end date in the header
-  echo -e "# $date_header\n\n$priorities" > "$HOME/Documents/Brain dump/Päivän suunnittelu/${start_day}-${end_day} (useampi päivä).md"
+  echo -e "# $date_header\n\n$priorities" > "$HOME/Documents/Brain dump/Päivän suunnittelu/$(date -d "$start_day" "%-d.%-m.%Y").md"
 
   echo -e "${BOLD}${GREEN}Prioritization is ready and saved to Obsidian.${RESET}"